Revisions due to Jesse's code review.
