+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/decisiontree/ProbabilisticClassifier.scala
@@ -41,16 +41,7 @@ object ProbabilisticClassifier {
     // Override the Spray/JSON serialization for maps with integer keys, because these
     // don't work out-of-the-box.
     import spray.json.DefaultJsonProtocol.{ mapFormat => _, _ }
-    implicit val intMapFormat = new JsonFormat[Map[Int, Int]] {
-      override def write(map: Map[Int, Int]): JsValue = {
-        map.toSeq.toJson
-      }
-      override def read(json: JsValue): Map[Int, Int] = json match {
-        case value: JsArray => value.convertTo[Seq[(Int, Int)]].toMap
-        case _ => ???
-      }
-    }
-    implicit val intSeqMapFormat = indexedSeqFormat[Map[Int, Int]]
+    import org.allenai.nlpstack.parse.poly.decisiontree.DecisionTree.intMapFormat
 
     implicit val decisionTreeFormat =
       jsonFormat4(DecisionTree.apply).pack("type" -> "DecisionTree")
